@@ -22,14 +22,14 @@
         }
     ],
     "require": {
-        "php": "^7.3",
-        "laravel/framework": "^7.0|^8.0"
+        "php": "^8.0",
+        "laravel/framework": "^8.0"
     },
     "require-dev": {
         "kint-php/kint": "^3.3",
-        "mockery/mockery": "^1.3",
-        "orchestra/testbench": "^5.1|^6.0",
-        "phpunit/phpunit": "^8.5"
+        "mockery/mockery": "^1.4.3",
+        "orchestra/testbench": "^6.0",
+        "phpunit/phpunit": "^9.3"
     },
     "autoload": {
         "psr-4": {
@@ -47,6 +47,8 @@
     "config": {
         "sort-packages": true
     },
+    "minimum-stability": "dev",
+    "prefer-stable": true,
     "extra": {
         "laravel": {
             "providers": [
@@ -70,7 +72,5 @@
             "@test-lowest",
             "@test-stable"
         ]
-    },
-    "minimum-stability": "dev",
-    "prefer-stable": true
+    }
 }
