@@ -10,12 +10,7 @@ use SebastiaanLuca\AutoMorphMap\Commands\ClearCachedMorphMap;
 
 class AutoMorphMapServiceProvider extends ServiceProvider
 {
-    /**
-     * Register the application services.
-     *
-     * @return void
-     */
-    public function register() : void
+    public function register(): void
     {
         $this->mergeConfigFrom(
             $this->getConfigurationPath(),
@@ -28,48 +23,31 @@ class AutoMorphMapServiceProvider extends ServiceProvider
         );
     }
 
-    /**
-     * Bootstrap the application services.
-     *
-     * @return void
-     */
-    public function boot() : void
+    public function boot(): void
     {
         $this->registerPublishableResources();
 
         app(Mapper::class)->map();
     }
 
-    /**
-     * @return void
-     */
-    private function registerPublishableResources() : void
+    private function registerPublishableResources(): void
     {
         $this->publishes([
-            $this->getConfigurationPath() => config_path($this->getShortPackageName() . '.php'),
-        ], $this->getPackageName() . ' (configuration)');
+            $this->getConfigurationPath() => config_path($this->getShortPackageName().'.php'),
+        ], $this->getPackageName().' (configuration)');
     }
 
-    /**
-     * @return string
-     */
-    private function getConfigurationPath() : string
+    private function getConfigurationPath(): string
     {
-        return __DIR__ . '/../config/' . $this->getShortPackageName() . '.php';
+        return __DIR__.'/../config/'.$this->getShortPackageName().'.php';
     }
 
-    /**
-     * @return string
-     */
-    private function getShortPackageName() : string
+    private function getShortPackageName(): string
     {
         return 'auto-morph-map';
     }
 
-    /**
-     * @return string
-     */
-    private function getPackageName() : string
+    private function getPackageName(): string
     {
         return 'laravel-auto-morph-map';
     }
